backend2/requirements.txt
Django==5.1.7
djangorestframework==3.15.2
celery==5.4.0
redis==5.2.1
channels==4.2.0
psycopg2-binary==2.9.10
kafka-python==2.1.1
django-cors-headers==4.7.0
django-filter==23.5
requests==2.32.3